+++ packages/api/src/city/city-fetch.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { CityResponseDto } from 'src/city/types/city-response.type';
 import { ExternalApiException } from 'src/common/errors/external-api.error';
-import { CityFetch } from './interfaces/city-fetch.interface';
+import { CityFetch } from './abstractions/city-fetch.abstract';
 
 @Injectable()
 export class CityFetchService implements CityFetch {
