+++ packages/api/src/common/abstractions/weather-handler.abstract.ts
@@ -6,7 +6,7 @@ export abstract class ProviderHandler<Response> {
 
   async handle(city: string): Promise<Response> {
     try {
-      const result = await this.fetch(city);
+      const result = await this.process(city);
       return result;
     } catch (err) {
       if (this.next && (err instanceof ExternalApiException || err instanceof CityNotFoundException)) {
@@ -16,7 +16,7 @@ export abstract class ProviderHandler<Response> {
     }
   }
 
-  abstract fetch(city: string): Promise<Response>;
+  abstract process(city: string): Promise<Response>;
   abstract get providerName(): string;
 
   setNext(handler: ProviderHandler<Response>) {
