+++ packages/api/src/notifications/notifications.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import { NotificationsServiceInterface } from 'src/notifications/interfaces/notifications-service.interface';
-import { NotificationsSender, NotificationsSenderToken } from './interfaces/notifications-sender.interface';
+import { NotificationsServiceInterface } from 'src/notifications/abstractions/notifications-service.abstract';
+import { NotificationsSender, NotificationsSenderToken } from './abstractions/notifications-sender.interface';
 import { MailModule } from './mail/mail.module';
 import { MailSender } from './mail/services/mail-sender.service';
 import { NotificationsService } from './notifications.service';
