+++ packages/api/src/subscriptions/services/subscription.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model, RootFilterQuery } from 'mongoose';
-import { ServiceSubscriptionRepository } from 'src/subscriptions/interfaces/subscription-repository.interface';
+import { ServiceSubscriptionRepository } from 'src/subscriptions/abstractions/subscription-repository.abstract';
 import { HOUR } from 'src/utils/time-units';
 import { Subscription, SubscriptionWithId } from '../../database/schemas/subscription.schema';
 import { CreateSubscriptionDto } from '../dtos/create-subscription.dto';
