+++ packages/api/src/subscriptions/subscription.module.ts
@@ -3,8 +3,8 @@ import { MongooseModule } from '@nestjs/mongoose';
 import { CityModule } from 'src/city/city.module';
 import { NotificationsModule } from 'src/notifications/notifications.module';
 import { Subscription, SubscriptionSchema } from '../database/schemas/subscription.schema';
-import { ControllerSubscriptionService, FindSubscriptionService } from './interfaces/subcription-service.interface';
-import { GroupSubscriptionRepository, ServiceSubscriptionRepository } from './interfaces/subscription-repository.interface';
+import { ControllerSubscriptionService, FindSubscriptionService } from './abstractions/subcription-service.abstract';
+import { GroupSubscriptionRepository, ServiceSubscriptionRepository } from './abstractions/subscription-repository.abstract';
 import { SubscriptionRepository } from './services/subscription.repository';
 import { SubscriptionService } from './services/subscription.service';
 import { SubscriptionController } from './subscription.controller';
