+++ packages/api/src/weather/abstractions/weather-fetch.abstract.ts
