+++ packages/api/src/weather/weather.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
+import { WeatherServiceInterface } from './abstractions/current-weather.abstract';
+import { WeatherFetch } from './abstractions/weather-fetch.abstract';
 import { CurrentOpenWeatherHandler } from './handlers/weather-openweather.handler';
 import { CurrentWeatherApiHandler } from './handlers/weather-weatherapi.handler';
-import { WeatherServiceInterface } from './interfaces/current-weather.interface';
-import { WeatherFetch } from './interfaces/weather-fetch.interface';
 import { WeatherFetchService } from './services/weather-fetch.service';
 import { WeatherService } from './services/weather.service';
 import { WeatherController } from './weather.controller';
