+++ packages/api/test/integration/subscription.integration.spec.ts
@@ -4,17 +4,17 @@ import { getModelToken, MongooseModule } from '@nestjs/mongoose';
 import { Test } from '@nestjs/testing';
 import { useContainer } from 'class-validator';
 import mongoose, { Model, Types } from 'mongoose';
+import { CityFetch } from 'src/city/abstractions/city-fetch.abstract';
 import { CityModule } from 'src/city/city.module';
-import { CityFetch } from 'src/city/interfaces/city-fetch.interface';
 import { CityWeatherApiFetchDto } from 'src/city/types/city-response.type';
 import { ExternalApiException } from 'src/common/errors/external-api.error';
 import { DatabaseExceptionFilter } from 'src/common/filters/database-exception.filter';
 import { DatabaseModule } from 'src/database/database.module';
 import { Subscription, SubscriptionSchema } from 'src/database/schemas/subscription.schema';
+import { NotificationsServiceInterface } from 'src/notifications/abstractions/notifications-service.abstract';
 import { NotificationsFrequencies } from 'src/notifications/constants/enums/notification-frequencies.enum';
 import { NotificationSubjects } from 'src/notifications/constants/enums/notification-subjects.enum';
 import { NotificationType } from 'src/notifications/constants/enums/notification-type.enum';
-import { NotificationsServiceInterface } from 'src/notifications/interfaces/notifications-service.interface';
 import { NotificationsModule } from 'src/notifications/notifications.module';
 import { CreateSubscriptionDto } from 'src/subscriptions/dtos/create-subscription.dto';
 import { SubscriptionRepository } from 'src/subscriptions/services/subscription.repository';
