+++ packages/api/test/integration/weather.integration.spec.ts
@@ -3,8 +3,8 @@ import { ConfigModule } from '@nestjs/config';
 import { Test } from '@nestjs/testing';
 import { CityNotFoundException } from 'src/common/errors/city-not-found.error';
 import { ExternalApiException } from 'src/common/errors/external-api.error';
+import { WeatherFetch } from 'src/weather/abstractions/weather-fetch.abstract';
 import { CurrentWeatherResponseDto } from 'src/weather/dtos/current-weather-response.dto';
-import { WeatherFetch } from 'src/weather/interfaces/weather-fetch.interface';
 import { CurrentOpenWeatherFetchDto, CurrentWeatherApiFetchDto } from 'src/weather/types/current-weather-api.type';
 import { WeatherModule } from 'src/weather/weather.module';
 import * as request from 'supertest';
